requirements.txt
motor==3.4.0
fastapi
uvicorn
python-dotenv